student layout
16-08-2023

+++ database/migrations/2023_08_06_980000_create_student_payments_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('academic_year_id');
             $table->unsignedBigInteger('student_id');
             $table->unsignedBigInteger('admission_id');
-            $table->integer('total_amount');
+            $table->double('total_amount');
             $table->tinyInteger('status')->nullable()->default('0')->comment('0-Not Paid,1-Paid,2-Cancel');
             $table->timestamps();
             $table->foreign('academic_year_id')->references('id')->on('academic_years')->onDelete('cascade');
